+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - GFX: Correct refreshing image when raw editing DefineExternalImage/2, DefineSubImage
 - GFX: DefineExternalImage/2, DefineSubImage disallow not working replace button in favor of raw editing
 - [#1795] AS3 P-code - optional (default parameter values) saving
+- [#1785] AS1/2 try..catch block in for..in
 
 ## [16.0.4] - 2022-11-03
 ### Fixed
@@ -2484,6 +2485,7 @@ All notable changes to this project will be documented in this file.
 [#1857]: https://www.free-decompiler.com/flash/issues/1857
 [#1818]: https://www.free-decompiler.com/flash/issues/1818
 [#1795]: https://www.free-decompiler.com/flash/issues/1795
+[#1785]: https://www.free-decompiler.com/flash/issues/1785
 [#1860]: https://www.free-decompiler.com/flash/issues/1860
 [#1782]: https://www.free-decompiler.com/flash/issues/1782
 [#1679]: https://www.free-decompiler.com/flash/issues/1679